Synchronisation rover et panneaux solaires

+++ Assets/Gameplay/Rover/RoverBattery.cs
@@ -2,18 +2,22 @@
 using UnityEngine.Networking;
 using System.Collections;
 
-public class RoverBattery : NetworkBehaviour {
+public class RoverBattery : NetworkBehaviour
+{
     public bool battery;
     public Quest qst;
     public RoverDisplace rv_disp;
     public Light roverLight;
-	// Use this for initialization
-	void Start () {
+    // Use this for initialization
+    void Start()
+    {
+        if (!isServer)
+            transform.rotation = GameObject.Find("NetworkManager").GetComponent<NetworkManager>().spawnPrefabs[1].transform.rotation;
         roverLight = gameObject.GetComponent<Light>();
         if (!battery)
             roverLight.enabled = false;
-	}
-	
+    }
+
 
 
     void OnTriggerEnter(Collider other)
@@ -23,19 +27,20 @@ public class RoverBattery : NetworkBehaviour {
 
     void OnTriggerStay(Collider other)
     {
-        if (Input.GetKeyDown(KeyCode.E)&& qst.item_list.quest.Count > 0 && qst.item_list.quest[qst.index].id == "battery")
-            {
-                PlayerSync player = GameObject.Find("PlayerContain").GetComponent<PlayerContain>().player_obj.GetComponent<PlayerSync>();
-                if (isServer)
-                    RpcSyncRover();
-                else
-                    player.CmdSyncRover();
-                roverLight.enabled = true;
-                qst.item_list.DeleteItem("battery");
-                qst.index = 0;
-                qst.UpdateName();
-                rv_disp.battery = true;
-            }
+        
+        if (Input.GetKeyDown(KeyCode.E) && qst != null && qst.item_list.quest.Count > 0 && qst.item_list.quest[qst.index].id == "battery")
+        {
+            PlayerSync player = GameObject.Find("PlayerContain").GetComponent<PlayerContain>().player_obj.GetComponent<PlayerSync>();
+            if (player.isServer)
+                RpcSyncRover();
+            else
+                player.CmdSyncRover();
+            roverLight.enabled = true;
+            qst.item_list.DeleteItem("battery");
+            qst.index = 0;
+            qst.UpdateName();
+            rv_disp.battery = true;
+        }
     }
 
     [ClientRpc]
